slider updated to rows
