+++ frontend/src/styles/index.scss
@@ -50,7 +50,7 @@
 @import "modules/tags";
 
 // COMPONENTS
-@import "components/cardSlider";
+@import "components/cardRow";
 @import "components/career-map";
 @import "components/career-pathways";
 @import "components/ctaBanner";
