+++ frontend/src/styles/pages/_home-page.scss
@@ -55,7 +55,7 @@
     }
   }
 
-  .card-slider {
+  .card-row {
     &:nth-child(2n) {
       background-color: $base-cool;
